requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy==2.0.40
pydantic>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
python-multipart==0.0.20
python-dotenv>=1.0.1